test/inspector/debugger/wasm-stepping-expected.txt

Tests stepping through wasm scripts by byte offsets

Running test: test
Setting up global instance variable.
Got wasm script: wasm://wasm/ce1a7b8a
Setting breakpoint on offset 72 (should be propagated to 73, the offset of the call), url wasm://wasm/ce1a7b8a
{
    columnNumber : 73
    lineNumber : 0
    scriptId : <scriptId>
}
Script wasm://wasm/ce1a7b8a byte offset 73: Wasm opcode 0x10 (kExprCallFunction)
Debugger.stepInto called
Script wasm://wasm/ce1a7b8a byte offset 52: Wasm opcode 0x01 (kExprNop)
Debugger.stepOver called
Script wasm://wasm/ce1a7b8a byte offset 53: Wasm opcode 0x01 (kExprNop)
Debugger.stepOut called
Script wasm://wasm/ce1a7b8a byte offset 75: Wasm opcode 0x0c (kExprBr)
Debugger.stepOut called
Script wasm://wasm/ce1a7b8a byte offset 73: Wasm opcode 0x10 (kExprCallFunction)
Debugger.stepOver called
Script wasm://wasm/ce1a7b8a byte offset 75: Wasm opcode 0x0c (kExprBr)
Debugger.stepInto called
Script wasm://wasm/ce1a7b8a byte offset 59: Wasm opcode 0x20 (kExprLocalGet)
Debugger.resume called
Script wasm://wasm/ce1a7b8a byte offset 73: Wasm opcode 0x10 (kExprCallFunction)
Debugger.stepInto called
Script wasm://wasm/ce1a7b8a byte offset 52: Wasm opcode 0x01 (kExprNop)
Debugger.stepOut called
Script wasm://wasm/ce1a7b8a byte offset 75: Wasm opcode 0x0c (kExprBr)
Debugger.stepInto called
Script wasm://wasm/ce1a7b8a byte offset 59: Wasm opcode 0x20 (kExprLocalGet)
Debugger.stepInto called
Script wasm://wasm/ce1a7b8a byte offset 61: Wasm opcode 0x04 (kExprIf)
Debugger.stepInto called
Script wasm://wasm/ce1a7b8a byte offset 63: Wasm opcode 0x20 (kExprLocalGet)
Debugger.stepInto called
Script wasm://wasm/ce1a7b8a byte offset 65: Wasm opcode 0x41 (kExprI32Const)
Debugger.stepInto called
Script wasm://wasm/ce1a7b8a byte offset 67: Wasm opcode 0x6b (kExprI32Sub)
Debugger.stepInto called
Script wasm://wasm/ce1a7b8a byte offset 68: Wasm opcode 0x21 (kExprLocalSet)
Debugger.stepInto called
Script wasm://wasm/ce1a7b8a byte offset 70: Wasm opcode 0x41 (kExprI32Const)
Debugger.stepInto called
Script wasm://wasm/ce1a7b8a byte offset 73: Wasm opcode 0x10 (kExprCallFunction)
Debugger.stepInto called
Script wasm://wasm/ce1a7b8a byte offset 52: Wasm opcode 0x01 (kExprNop)
Debugger.stepInto called
Script wasm://wasm/ce1a7b8a byte offset 53: Wasm opcode 0x01 (kExprNop)
Debugger.stepInto called
Script wasm://wasm/ce1a7b8a byte offset 54: Wasm opcode 0x0b (kExprEnd)
Debugger.stepInto called
Script wasm://wasm/ce1a7b8a byte offset 75: Wasm opcode 0x0c (kExprBr)
Debugger.stepInto called
Script wasm://wasm/ce1a7b8a byte offset 59: Wasm opcode 0x20 (kExprLocalGet)
Debugger.resume called
exports.main returned!
Test stepping over a recursive call
Setting breakpoint on the recursive call instruction @+93, url wasm://wasm/ce1a7b8a
{
    columnNumber : 93
    lineNumber : 0
    scriptId : <scriptId>
}
Script wasm://wasm/ce1a7b8a byte offset 93: Wasm opcode 0x10 (kExprCallFunction)
Removing breakpoint
Debugger.stepOver called
Script wasm://wasm/ce1a7b8a byte offset 95: Wasm opcode 0x20 (kExprLocalGet)
Debugger.resume called
